Simulated annealing implementation
This is a C++ implementation of the simulated annealing metaheuristic on
the Job shop or flow shop problem. The program takes a list of jobs and
applies the simulated annealing algorithm to find the best schedule for
an optimal makespan. The program allows the user to enter his own set of
jobs in or will randomly generate a list itself.